fix linux build without vulkan
